% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic_incidents_wrangling.R
\name{construction_detours_wrangling}
\alias{construction_detours_wrangling}
\title{construction_detours_wrangling cleans the traffic incidents raw data}
\usage{
construction_detours_wrangling(data)
}
\arguments{
\item{construction_detours_wrangling}{Construction detours raw data}
}
\value{
data frame
}
\description{
construction_detours_wrangling cleans the traffic incidents raw data
}
\details{
This function cleans the column names and converts the date columns to date time format
}
