% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{download_data gets raw data from the YYC Open Data Portal, or really any provided url}
\usage{
download_data(url, destfile)
}
\arguments{
\item{url}{Persistent url to the data}

\item{destfile}{Destination to save downloaded file, including the file name}
}
\value{
csv file
}
\description{
download_data gets raw data from the YYC Open Data Portal, or really any provided url
}
\examples{
traffic_incidents_url <- "https://data.calgary.ca/resource/35ra-9556.csv"

# Running the functions
download_data(url = traffic_incidents_url, destfile = "00_data/traffic_incidents.csv")

}
