% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic_incidents_wrangling.R
\name{traffic_incidents_wrangling}
\alias{traffic_incidents_wrangling}
\title{traffic_incidents_wrangling cleans the traffic incidents raw data}
\usage{
traffic_incidents_wrangling(data)
}
\arguments{
\item{traffic_incidents_raw_data}{Traffic incidents raw data}
}
\value{
data frame
}
\description{
traffic_incidents_wrangling cleans the traffic incidents raw data
}
\details{
This function cleans the column names and converts the date columns to date time format
}
